Add: Feed Delete, update lock API 통신 비즈니스 로직 완성
- TODO: update lock api response 업데이트 시 반영

+++ app/src/main/java/com/sgs/devcamp2/flametalk_android/network/response/feed/Feed.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 @Keep
 data class Feed(
     @SerializedName("feedId")
-    val feedId: Int,
+    val feedId: Long,
     @SerializedName("imageUrl")
     val imageUrl: String,
     @SerializedName("isBackground")
